example server moved to seperate file
